feat: Create a sample TODO webapp
Create a sample TODO app which allows defining tasks and their categories.
The existing 2 App Services Web Apps (web + api) will be dropped in favor of using this single web app to make it easier to check the entirety of the project, change it and run it (locally and in Azure).

noissue